feat: move Bids into MyStore, remove Partners
