+++ webapp/src/components/AssetPage/BidList/BidList.container.ts
@@ -1,9 +1,9 @@
-import { MapStateProps, MapDispatchProps, MapDispatch } from './Bids.types'
+import { MapStateProps, MapDispatchProps, MapDispatch } from './BidList.types'
 import { RootState } from '../../../modules/reducer'
 import { getNFTBids } from '../../../modules/ui/nft/bid/selectors'
 import { fetchBidsByNFTRequest } from '../../../modules/bid/actions'
 import { connect } from 'react-redux'
-import Bids from './Bids'
+import BidList from './BidList'
 
 const mapState = (state: RootState): MapStateProps => ({
   bids: getNFTBids(state)
@@ -13,4 +13,4 @@ const mapDispatch = (dispatch: MapDispatch): MapDispatchProps => ({
   onFetchBids: nft => dispatch(fetchBidsByNFTRequest(nft))
 })
 
-export default connect(mapState, mapDispatch)(Bids)
+export default connect(mapState, mapDispatch)(BidList)
