+++ webapp/src/components/AssetPage/BidList/BidList.css
@@ -1,3 +1,3 @@
-.Bids .list {
+.BidList .list {
   margin-top: 13px;
 }
