+++ webapp/src/components/AssetPage/BidList/BidList.types.ts
