+++ webapp/src/components/AssetPage/BidList/index.ts
@@ -1,0 +1,2 @@
+import BidList from './BidList.container'
+export { BidList }
