+++ webapp/src/components/AssetPage/ENSDetail/ENSDetail.tsx
@@ -7,7 +7,7 @@ import { Owner } from '../Owner'
 import Price from '../Price'
 import Expiration from '../Expiration'
 import { Actions } from '../Actions'
-import { Bids } from '../Bids'
+import { BidList } from '../BidList'
 import { TransactionHistory } from '../TransactionHistory'
 import BaseDetail from '../BaseDetail'
 import { AssetImage } from '../../AssetImage'
@@ -29,7 +29,7 @@ const ENSDetail = ({ nft }: Props) => (
     }
     below={
       <>
-        <Bids nft={nft} />
+        <BidList nft={nft} />
         <TransactionHistory nft={nft} />
       </>
     }
