+++ webapp/src/components/AssetPage/ParcelDetail/ParcelDetail.tsx
@@ -9,7 +9,7 @@ import { Owner } from '../Owner'
 import Price from '../Price'
 import Expiration from '../Expiration'
 import { Actions } from '../Actions'
-import { Bids } from '../Bids'
+import { BidList } from '../BidList'
 import { TransactionHistory } from '../TransactionHistory'
 import { Coordinate } from '../../Coordinate'
 import { JumpIn } from '../JumpIn'
@@ -71,7 +71,7 @@ const ParcelDetail = ({ nft }: Props) => {
       }
       below={
         <>
-          <Bids nft={nft} />
+          <BidList nft={nft} />
           <TransactionHistory nft={nft} />
         </>
       }
