+++ webapp/src/components/Bids/Bids.types.ts
