+++ webapp/src/components/Bids/index.ts
