--- webapp/src/components/MyBidsPage/index.ts
@@ -1,2 +1,0 @@
-import MyBidsPage from './MyBidsPage.container'
-export { MyBidsPage }
