+++ webapp/src/components/Navigation/Navigation.types.ts
@@ -2,10 +2,7 @@ export enum NavigationTab {
   LANDS = 'lands',
   BROWSE = 'browse',
   COLLECTIBLES = 'collectibles',
-  PARTNERS = 'partners',
-  PARTNER = 'partner',
   MY_STORE = 'my_store',
-  MY_BIDS = 'my_bids',
   ACTIVITY = 'activity'
 }
