--- webapp/src/components/PartnersPage/PartnersPage.css
@@ -1,15 +1,0 @@
-.PartnersPage .Layout {
-  display: flex;
-}
-
-.PartnersPage .Layout .column {
-  height: 100%;
-}
-
-.PartnersPage .Row > .Column.right {
-  overflow: hidden;
-}
-
-.PartnersPage .Layout .Column.right > div {
-  align-self: start;
-}
