--- webapp/src/components/PartnersPage/index.ts
@@ -1,2 +1,0 @@
-import PartnersPage from './PartnersPage.container'
-export { PartnersPage }
