--- webapp/src/components/VendorStrip/index.ts
@@ -1,2 +1,0 @@
-import VendorStrip from './VendorStrip.container'
-export { VendorStrip }
