+++ webapp/src/modules/translation/locales/en.json
@@ -44,12 +44,6 @@
   "estate_detail": {
     "dissolved": "This Estate has been dissolved"
   },
-  "error_boundary": {
-    "message": "Oops! Something went wrong..."
-  },
-  "estate_detail": {
-    "dissolved": "This Estate has been dissolved"
-  },
   "vendors": {
     "decentraland": "Decentraland",
     "super_rare": "SuperRare",
@@ -97,6 +91,7 @@
     "collections": "Collections",
     "on_sale": "On Sale",
     "sales": "Sales",
+    "bids": "Bids",
     "store_settings": "Settings"
   },
   "vendor_menu": {
